Group queries and test getInsertPoint
